// Qanat.Web/src/app/shared/generated/model/water-account-simple-dto.ts
export class WaterAccountSimpleDto { 
    WaterAccountID?: number;
    GeographyID?: number;
    WaterAccountNumber?: number;
    WaterAccountName?: string;
    Notes?: string;
    UpdateDate?: string;
    WaterAccountPIN?: string;
    CreateDate?: string;
    ContactName?: string;
    ContactAddress?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}